changed all floats to doubles

+++ lib/element.h
@@ -6,7 +6,7 @@
 /*   By: yiwong <yiwong@student.42wolfsburg.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/15 18:36:55 by yiwong            #+#    #+#             */
-/*   Updated: 2024/01/04 17:30:26 by yiwong           ###   ########.fr       */
+/*   Updated: 2024/01/05 15:27:04 by yiwong           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@
 
 typedef struct s_ambience
 {
-	float		lighting;
+	double		lighting;
 	t_rgb		*colour;
 }				t_ambience;
 
@@ -33,14 +33,14 @@ typedef struct s_camera
 typedef struct s_light
 {
 	t_vector	*point;
-	float		brightness;
+	double		brightness;
 	t_rgb		*colour;
 }				t_light;
 
 typedef struct s_sphere
 {
 	t_vector	*centre;
-	float		diameter;
+	double		diameter;
 	t_rgb		*colour;
 }				t_sphere;
 
@@ -55,8 +55,8 @@ typedef struct s_cylinder
 {
 	t_vector	*centre;
 	t_vector	*axis;
-	float		diameter;
-	float		height;
+	double		diameter;
+	double		height;
 	t_rgb		*colour;
 }				t_cylinder;
