+++ src/math/math_utils.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   math_utils.h                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: Kekuhne <kekuehne@student.42wolfsburg.d    +#+  +:+       +#+        */
+/*   By: yiwong <yiwong@student.42wolfsburg.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/06 17:29:29 by Kekuhne           #+#    #+#             */
-/*   Updated: 2024/01/03 19:54:40 by Kekuhne          ###   ########.fr       */
+/*   Updated: 2024/01/05 15:27:04 by yiwong           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,14 +15,14 @@
 
 typedef struct s_vector 
 {
-	float x;
-	float y;
-	float z;
+	double x;
+	double y;
+	double z;
 } t_vector;
 
-float	ft_sqr(float n);
-void	set_vector_to(t_vector *set, float x, float y, float z);
-void	set_vector_to_single(t_vector *set, float n);
-float	vector_lenght(t_vector *vec);
-float	vector_normalize(t_vector *vec);
+double	ft_sqr(double n);
+void	set_vector_to(t_vector *set, double x, double y, double z);
+void	set_vector_to_single(t_vector *set, double n);
+double	vector_lenght(t_vector *vec);
+double	vector_normalize(t_vector *vec);
 #endif
